+++ dev/200104 bug when reading grid.R
@@ -1,0 +1,1 @@
+x <- importTxt("dev/data/grid.txt")